Extract Keywords: Refactor keywords library
